add attr for numberpicker

+++ numpad/src/main/java/hseify69/ir/numpad/helpers/Consts.java
@@ -2,8 +2,8 @@ package hseify69.ir.numpad.helpers;
 
 public class Consts {
 
-    public static final long initValue = 1;
-    public static final long initMaxValue = 10;
-    public static final long initMinValue = 1;
+    public static final int initValue = 1;
+    public static final int initMaxValue = 10;
+    public static final int initMinValue = 1;
 
 }
